+++ src/app/(providers)/(root)/[workspaceId]/_components/MemberList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MemberList';
